pyproject.toml
[project]
name = "mirror-docs"
version = "0.1.0"
description = "Mirrors HTML documentation to Markdown for AI integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lxml[html-clean]>=5.3.2",
    "markdownify>=1.1.0",
    "readability-lxml>=0.8.1",
    "beautifulsoup4>=4.12.0",
]

[project.scripts]
mirror-docs = "mirror_docs:main"
